added waitForExit functionality, moved from util to node due to import cycle error
